server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-session
flask-cors
python-dotenv
redis